requirements.txt
anthropic
python-dotenv
streamlit
ddgs
rich
